[rhobs/prod{mst,telemeter}]: Add separate Queriers for Rulers
This commit rolls out separate Queriers for Rulers and fully removes any
vestiges of stateless Rulers. It adds new sets of Thanos Query with the
same config as the regular ones, and modifies the Rulers to query the new
Queriers instead.

This is done for the following production envs, as they are linked,
* MST Production
* Telemeter Production
* For MST Metrics Fed Rulers, we only change the Query addr,
  to be pointed to new MST ruler queriers

This commit does not roll the split out for rhobsp02ue1 MST.

Signed-off-by: Saswata Mukherjee <samukher@redhat.com>