Remove ocm-quay memory alert
It's been triggering for days with neither any action nor consequence,
so it's clear that this alert is of no use.

Taking a look into the limits of the pods, they're set to 56Gb. Why this
alert exists then instead of setting the limit escapes to my
understanding. Being ocm-quay close to being decomissioned, it looks
that it is not that important anyway...

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>