Removed quay-mirror-org int stuck alert
It currently takes more than 9 hours in compare-tags mode. It is not
common that the integration runs frequently in this mode as it gets
stopped often due to deployments.

Since we know the behaviour of the integration and we have an alert that
will tell us if it fails to run, it doesn't make a lot of sense to
include quay-mirror-org in this alert, as it is always a no-op which
decreases our confidence in this alert.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>