feature: upgrade minor version of the stage database
In order to be able to upgrade the database to the latest major plus
minor version, we need to make sure first that the minor version of the
current database is the latest one. According to the "Upgrading the
PostgreSQL DB engine for Amazon RDS" document[1], we should be upgrading
our database to version "11.20".

[1]: https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_UpgradeDBInstance.PostgreSQL.html

RHCLOUD-26954