Promote hive a84d50cb1bda4b87714c57a1cf66b111a7cd6c12
https://github.com/openshift/hive/compare/212207882cc9304a61f117b7d349eaf037e6edb3...a84d50cb1bda4b87714c57a1cf66b111a7cd6c12

---
```
commit: 8d7c2a5c1b731147262ea3fd846557e5bf7bf23c
author: Eric Fried
e2e (AWS): Allow env var to override creds file

Resolve a long-standing TODO which allows AWS environment variables
`AWS_ACCESS_KEY_ID` and `AWS_SECRET_ACCESS_KEY` to be used when present,
even if a credentials file exists, as the gods intended.

hiveutil already uses the appropriate logic such that, if a creds file
is not supplied, it looks for the env vars. We just needed to fix it so
we don't pass that arg in at all if the env vars are set.

(This was an incidental requirement of HIVE-2234: the AWS credentials
from the hypershift installation were taking precedence for managed DNS,
which then bounced because the expected hosted zone is in a different
account.)

commit: 13d580c75377b60f6fedfac68b7a0a28893cd29f
author: Feilian Xie
Add bash script which allows running Hive components locally, simulating in-cluster environment

commit: 30afd4bb44d17903dac76bbb5a33510cd546ebb0
author: Eric Fried
MachinePool: hack name of vSphere default worker MachineSet

With recent installer changes to accommodate zones in vSphere, the name
of the default worker pool in OCP 4.13+ is now suffixed with a `-0` even
when only one zone is in play. Hive's day 2 MachinePool logic needs to
be able to generate the correct MachineSet name when talking to a 4.13+
cluster -- but also when talking to a 4.12- cluster. This commit adds in
a special case to discover and account for this difference where it
exists.

HIVE-2254

commit: fd1c8ed31c786ffc15ba36aabea2305355838a94
author: Eric Fried
MachinePool/GCP: support image override annotation

Add support for overriding the VM image used for GCP MachinePools via
the `hive.openshift.io/image-id-override` annotation, which was
previously only available for AWS.

This is (still) not part of the official API.

HIVE-2275

```