appsres05ue1: Create missing cluster pairings
Copied from appsres03ue1, since this cluster will replace it.

This follows:

* 514bc1d7ec53c42412b11840eb5019013ff48c78
* 030ca2c334ed277a3c8b4728d18a9f1a18ac4c8c

In backplane clusters (ROSA STS) we have added a
awsInfrastructureManagementAccount section to the requester side. This
is strictly not necessary, terraform-vpc-peerings works without this
because currently we force the two accounts to assume role from to be
equal and awsInfrastructureManagementAccounts is needed in the accepter
side, but it is certainly clearer how things work if it is provided here.

part of APPSRE-7925

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>