OCM-3321 | add delegation for ocm.devshift.org
We are currently using sdev.devshift.net for HyperShift OCM local developments.
In scope of the migration to AWS Managed Policies, only the devshift.org is
supported. Instead of asking for an exception to AWS which may be denied,
SRE-P suggested to migrate to use the .org domain instead, so we are adding with this
change the delegation for ocm.devshift.org  to the OCM AWS DEV account.

Meanwhile, we also changed the domain from *sdev* to *ocm* as sdev is
deprecated and being phased out.

Related: OCM-3321