Hive: Update hive-integration to master (explicit sha)
As part of testing the rollout of HIVE-2297 and OCM-3537, update just
one shard to the current master -- but explicitly by sha. We may do this
a couple more times, e.g. to validate
https://github.com/openshift/hive/pull/2105 once it merges. Eventually
we intend to put all stg/int shards back to tracking `master` by name.

HIVE-2297
OCM-3537